实现基本转换功能,对操作类型没有做完全支持 鉴于Google财经本身还有不少问题，2Gf暂停开发

+++ GfCsv.cs
@@ -27,7 +27,7 @@ namespace Jeebook._2Gf
         // 日期
         public DateTime Date;
         // 股数
-        public int Shares;
+        public double Shares;
         // 价格
         public double Price;
         // 现金价值
@@ -42,7 +42,7 @@ namespace Jeebook._2Gf
     {
         System.Collections.ArrayList arRecords = new System.Collections.ArrayList();
 
-        public void Add(string Symbol, string Name, GfType Type, DateTime Date, int Shares, double Price, double Commission, string Notes)
+        public void Add(string Symbol, string Name, GfType Type, DateTime Date, double Shares, double Price, double Commission, string Notes)
         {
             GfRecord rec = new GfRecord();
             rec.Symbol = Symbol;
